feat: quit quiz when navigated to other page
