--- Main/html/quiz/entry.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
-    <title>싱숑샹숑</title>
-</head>
-<body>
-    <div id="topFixedBar" class="top-fixed-bar gradient">
-        <nav>
-            <button class="ranking-icon">
-                <i class="fa-solid fa-ranking-star"></i>
-            </button>        
-            <a class="main-top-logo" href="/">
-                <img class="main-top-logo-image" alt="err.png">
-            </a>
-            <button class="search-icon">
-                <i class="fa-solid fa-magnifying-glass"></i>
-            </button>
-        </nav>       
-        <div class="progress-container">
-            <div class="progress" style="width: 0%;"></div>
-        </div>
-    </div>
-    <div id="mainContainer" class="main-container">
-        <article class="quiz-entry display-none">
-            <div class="introduction">
-                <h1>수험자 정보 입력</h1>
-                <p>평점 기록 및 경품 수령을 위한 어쩌구입니다.</p>
-            </div>
-            <form>
-                <label>
-                    닉네임
-                    <input type="input" id="nickname" min="6" required>
-                </label>
-                <label>
-                    이메일
-                    <input type="email" id="email" required>
-                </label>
-                <button type="submit">시작하기</button>
-            </form>
-        </article>
-        <article class="loading">
-            <i class="fa-solid fa-spin fa-spinner"></i>
-            잠시만 기다려주세요....
-        </article>
-    </div>
-</body>
-</html>
