+++ Main/src/api/dummy_data/posts.ts
@@ -47,7 +47,7 @@ export const dummyPosts: { [key: string]: PostBoardSectionData[] } =
                     {
                         landscape: generateDummyPost(
                             `${topCategoryName} 고인물 테스트 도전!`,
-                            "/quiz/entry.html?id=" +
+                            "/quiz/solve.html?id=" +
                                 encodeURIComponent(`nerd-${topCategoryName}`)
                         ),
                     },
@@ -56,7 +56,7 @@ export const dummyPosts: { [key: string]: PostBoardSectionData[] } =
                         portraits: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9].map((j) =>
                             generateDummyPost(
                                 `${i} 모의고사 ${j}탄`,
-                                `/quiz/entry.html?id=` +
+                                `/quiz/solve.html?id=` +
                                     encodeURIComponent(
                                         `quiz-${topCategoryName}-${i}-${j}`
                                     )
