+++ Main/src/images.d.ts
@@ -2,3 +2,4 @@ declare module "*.jpg";
 declare module "*.png";
 declare module "*.jpeg";
 declare module "*.gif";
+declare module "*.html";
