+++ Main/src/page/quiz/solve.ts
@@ -1,105 +1,46 @@
+import { dom, library } from "@fortawesome/fontawesome-svg-core";
 import "../../../styles/quiz";
-import { QuizSession } from "../../api/quiz_session";
-import initShareButton from "../../initShare";
-import {
-    addAnswerSubmitListener,
-    displayCorrectnessAnimation,
-    displayProblem,
-    initQuizSolveUI,
-    setHelpMeFriendsEventHandler,
-    updateProgress,
-    updateShareProblem,
-} from "../../quiz-solve-ui";
+import { QuizApiClient } from "../../api/quiz";
 import { InitTopNav } from "../../top_bottom_animation";
+import { faSpinner } from "@fortawesome/free-solid-svg-icons";
+import initSolvePage from "../../solvePage";
 
-initQuizSolveUI();
 InitTopNav();
 
-const sessionId =
-    new URLSearchParams(location.search.substring(1)).get("session") ?? "";
-const session = new QuizSession(sessionId);
-let shared = false;
-const setShareData = initShareButton(() => (shared = true));
+// FontAwesome 아이콘 렌더링
+library.add(faSpinner);
+dom.i2svg({ node: document.querySelector("article.loading")! });
 
-(async () => {
-    updateProgress(0);
-    const goResult = () =>
-        (location.href =
-            "/quiz/result.html?session=" + encodeURIComponent(sessionId));
-    const sessionInfo = await session.sessionInfo();
-
-    const renewProblem = async () => {
-        const problem = await session.currentProblem();
-        console.log(problem);
-        if (problem === null) {
-            return goResult();
-        }
-
-        displayProblem(
-            document.querySelector("article")!,
-            problem,
-            problem.index
-        );
-        updateShareProblem(
-            document.querySelector(".help-me .problem-box")!,
-            problem,
-            problem.index
-        );
-        setShareData({
-            text: "친구들야, 도와줘!",
-        });
-
-        if (!sessionInfo.isNerdTest) {
-            updateProgress(
-                ((problem.index - 1) / sessionInfo.totalProblemCount!) * 100,
-                `${problem.index}/${sessionInfo.totalProblemCount!}`
-            );
-            setHelpMeFriendsEventHandler({
-                onEnabled: () => {
-                    shared = false;
-                },
-                beforeDisable: () => shared,
-            });
-        } else {
-            setHelpMeFriendsEventHandler({
-                onDisabled: () => session.getStopWatch().resume(),
-                onEnabled: () => {
-                    session.getStopWatch().pause();
-                    shared = false;
-                },
-                beforeDisable: () => shared,
-            });
+const params = new URLSearchParams(location.search.substring(1));
+const quizId = params.get("id");
+if (quizId === null || quizId === " ") {
+    alert("오류");
+} else {
+    QuizApiClient.isNerdTest(quizId).then((isNerdTest) => {
+        if (isNerdTest) {
             document
-                .querySelector(".timer-paused")
+                .querySelector("article.loading")
+                ?.classList.add("display-none");
+            document
+                .querySelector("article.display-none")
                 ?.classList.remove("display-none");
+            document
+                .querySelector("form")
+                ?.addEventListener("submit", (evt) => {
+                    evt.preventDefault();
+                    QuizApiClient.startNerdQuiz(quizId, {
+                        nickname: (
+                            document.querySelector(
+                                "#nickname"
+                            ) as HTMLInputElement
+                        ).value,
+                        email: (
+                            document.querySelector("#email") as HTMLInputElement
+                        ).value,
+                    }).then(initSolvePage);
+                });
+        } else {
+            QuizApiClient.startQuiz(quizId).then(initSolvePage);
         }
-    };
-
-    addAnswerSubmitListener(async (answer) => {
-        const correct = await session.submit(answer);
-        displayCorrectnessAnimation(correct);
-
-        renewProblem();
-        console.log(answer);
     });
-
-    if (sessionInfo.isNerdTest) {
-        setInterval(() => {
-            const elapsed = session.getStopWatch().elapsed();
-            const totalTime = 1000 * 60 * 5;
-            const percentage = (elapsed / totalTime) * 100;
-            if (percentage >= 100) return goResult();
-
-            const leftTime = totalTime - elapsed;
-            updateProgress(
-                percentage,
-                `${Math.floor(leftTime / 1000 / 60)}:${(
-                    Math.floor(leftTime / 1000) % 60
-                )
-                    .toString()
-                    .padStart(2, "0")}`
-            );
-        }, 500);
-    }
-    renewProblem();
-})();
+}
