+++ Main/src/solvePage.html
@@ -1,0 +1,43 @@
+
+<div id="topFixedBar" class="top-fixed-bar gradient">
+    <nav>
+        <button class="ranking-icon">
+            <i class="fa-solid fa-ranking-star"></i>
+        </button>        
+        <a class="main-top-logo" href="/">
+            <img class="main-top-logo-image" alt="err.png">
+        </a>
+        <button class="search-icon">
+            <i class="fa-solid fa-magnifying-glass"></i>
+        </button>
+    </nav>
+    <div class="progress-container">
+        <div class="progress" style="width: 0%;"></div>
+        <div class="progress-text"></div>
+    </div>
+</div>
+<div id="mainContainer" class="main-container">
+    <article class="">
+        <i class="fa-solid fa-spin fa-spinner"></i>
+    </article>
+    <article class="help-me">
+        <div class="timer-paused display-none">타이머가 일시정지 되었습니다.</div>
+        <h1>친구들야, 도와줘!</h1>
+        <div class="problem-box">
+            </div>
+            <div class="buttons">
+            <button class="share-web-share">친구한테 물어보기</button>
+            <button class="share-twitter">트친한테 물어보기</button>
+            <button class="share-kakao">실친한테 물어보기</button>
+            <button class="continue">이어서 풀기</button>
+            </div>
+    </article>
+</div>
+<div class="correctness-effect display-none">
+    <div class="ok">
+        <i class="fa-solid fa-check"></i>
+    </div>
+    <div class="fail">
+        <i class="fa-solid fa-xmark"></i>
+    </div>
+</div>
