+++ package.json
@@ -11,6 +11,7 @@
   "devDependencies": {
     "css-loader": "^6.8.1",
     "file-loader": "^6.2.0",
+    "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.6.0",
     "mini-css-extract-plugin": "^2.7.6",
     "sass": "^1.69.7",
